remove brig proxy
`brig proxy` is identical to `brig dashboard --open-dashboard=false`.
